Add C++11 allocator
SICMAllocator version is chosen at compile time using __cplusplus
