+++ include/low/public/CMakeLists.txt
@@ -1,6 +1,8 @@
 set(SICM_PUBLIC_HEADERS
   sicm_low.h
   sicm.hpp
+  sicm-03.hpp
+  sicm-11.hpp
 )
 
 install(FILES ${SICM_PUBLIC_HEADERS}
